Find intersection between line (segment) and plane. Use double for distances

+++ Generic/Edge.cs
@@ -23,7 +23,7 @@ namespace Library.Generic
 
         public float Dz() => p2.z - p1.z;
 
-        public float Length() => (p2 - p1).Magnitude;
+        public double Length() => (p2 - p1).Magnitude;
 
         public static bool operator ==(Edge a, Edge b) => a.p1 == b.p1 && a.p2 == b.p2 || a.p1 == b.p2 && a.p2 == b.p1;
